Fix: Fix bug where some sizes wouldn't display correctly
For some reason, when I was using 100% as the basis for square
sizes, it would glitch out when I used numbers that evenly went
into 100 and would incorrectly distribute the pixels.

I am now using 99% as the max size of the canvas, which seems to
have fixed the problem. It's not a super elegant solution, but
it works for now.